break out InstanceSerializer
