+++ api/serializers/__init__.py
@@ -16,3 +16,4 @@ from .paginated_application_serializer import PaginatedApplicationSerializer
 from .application_bookmark_serializer import ApplicationBookmarkSerializer
 from .application_score_serializer import ApplicationScoreSerializer
 from .credential_serializer import CredentialSerializer
+from .instance_serializer import InstanceSerializer
